move files under separate packages, add backend util functions
