+++ src/web/backend/playlist/handlers.go
@@ -1,222 +1,33 @@
-package backend
+package playlist
 
 import (
 	"encoding/json"
-	"fmt"
-	"log/slog"
-	"math/rand/v2"
 	"net/http"
+	"fmt"
 	"os"
+	"log/slog"
 	"path/filepath"
-	"regexp"
-	"strings"
+	"math/rand/v2"
 	"time"
 
-	"explo/src/discovery"
 	"explo/src/util"
-	"explo/src/web"
 	"explo/src/web/backend/defs"
+	"explo/src/web"
 
 	"golang.org/x/text/cases"
 	"golang.org/x/text/language"
 )
 
-// CustomPlaylist holds the metadata for a user-imported playlist.
-type CustomPlaylist struct {
-	ID              string    `json:"id"`
-	Name            string    `json:"name"`
-	Source          string    `json:"source"`                     // "listenbrainz" | "apple_music" | "spotify"
-	SourceURL       string    `json:"source_url,omitempty"`       // original URL for dedup + refresh
-	LBMBID          string    `json:"lb_mbid,omitempty"`          // ListenBrainz MBID (backward compat)
-	ArtworkURL      string    `json:"artwork_url,omitempty"`      // playlist cover image (Apple Music)
-	ArtworkUploaded bool      `json:"artwork_uploaded,omitempty"` // true after artwork has been pushed to the music app
-	RefreshDays     int       `json:"refresh_days"`
-	ColorIndex      int       `json:"color_index"`
-	LastFetched     time.Time `json:"last_fetched"`
-}
-
-// CustomPlaylistArtworkPath returns the local file path where a playlist's
-// artwork is cached (regardless of whether the file exists).
-func CustomPlaylistArtworkPath(cfgDir, id string) string {
-	return filepath.Join(cfgDir, "cache", "playlist_artwork", id+".jpg")
-}
-
-// GetCustomPlaylist looks up a custom playlist by ID. Returns nil if not found.
-func GetCustomPlaylist(cfgDir, id string) *CustomPlaylist {
-	for _, p := range loadCustomPlaylists(cfgDir) {
-		if p.ID == id {
-			cp := p
-			return &cp
-		}
-	}
-	return nil
-}
-
-// MarkCustomPlaylistArtworkUploaded sets ArtworkUploaded=true and persists.
-func MarkCustomPlaylistArtworkUploaded(cfgDir, id string) error {
-	playlists := loadCustomPlaylists(cfgDir)
-	for i := range playlists {
-		if playlists[i].ID == id {
-			if playlists[i].ArtworkUploaded {
-				return nil
-			}
-			playlists[i].ArtworkUploaded = true
-			return saveCustomPlaylists(cfgDir, playlists)
-		}
-	}
-	return nil
-}
-
-var lbMBIDRe = regexp.MustCompile(`[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}`)
-
-var appleMusicURLRe = regexp.MustCompile(
-	`^https?://music\.apple\.com/[a-z]{2}/playlist/[^/]+/(pl\.[a-zA-Z0-9-]+)`,
-)
-
-// extractAppleMusicID pulls the playlist ID (pl.xxx) from an Apple Music URL.
-func extractAppleMusicID(raw string) (string, error) {
-	raw = strings.TrimSpace(raw)
-	m := appleMusicURLRe.FindStringSubmatch(raw)
-	if len(m) < 2 {
-		return "", fmt.Errorf("not a valid Apple Music playlist URL")
-	}
-	return m[1], nil
-}
-
-// extractLBMBID pulls the playlist UUID out of a ListenBrainz playlist URL or bare MBID string.
-func extractLBMBID(raw string) (string, error) {
-	raw = strings.TrimSpace(raw)
-	m := lbMBIDRe.FindString(raw)
-	if m == "" {
-		return "", fmt.Errorf("no ListenBrainz playlist UUID found in %q", raw)
-	}
-	return m, nil
-}
-
-func customPlaylistsPath(cfgDir string) string {
-	return filepath.Join(cfgDir, "custom-playlists.json")
-}
-
-// customEnvPrefix converts a playlist name like "Today's Hits"
-// to an env-var prefix like "CUSTOM_TODAYS_HITS".
-// Non-alphanumeric characters are collapsed into underscores.
-func customEnvPrefix(name string) string {
-	var b strings.Builder
-	prevUnderscore := true // start true so leading separators are skipped
-	for _, r := range strings.ToUpper(name) {
-		if (r >= 'A' && r <= 'Z') || (r >= '0' && r <= '9') {
-			b.WriteRune(r)
-			prevUnderscore = false
-		} else if !prevUnderscore {
-			b.WriteRune('_')
-			prevUnderscore = true
-		}
-	}
-	return "CUSTOM_" + strings.TrimRight(b.String(), "_")
-}
-
-
-func loadCustomPlaylists(cfgDir string) []CustomPlaylist {
-	data, err := os.ReadFile(customPlaylistsPath(cfgDir))
-	if err != nil {
-		return nil
-	}
-	var out []CustomPlaylist
-	if err := json.Unmarshal(data, &out); err != nil {
-		slog.Warn("custom-playlists: failed to parse metadata", "err", err)
-		return nil
-	}
-	return out
-}
-
-func saveCustomPlaylists(cfgDir string, playlists []CustomPlaylist) error {
-	raw, err := json.MarshalIndent(playlists, "", "  ")
-	if err != nil {
-		return err
-	}
-	return os.WriteFile(customPlaylistsPath(cfgDir), raw, 0644)
-}
-
-// FetchResult is the uniform return type for fetching playlist data from any source.
-type FetchResult struct {
-	Name       string
-	ArtworkURL string
-	Tracks     []PlaylistTrack
-}
-
-// fetchCustomPlaylistTracks dispatches to the appropriate source fetcher.
-// This is the single point where source-specific logic lives for fetching.
-func fetchCustomPlaylistTracks(p CustomPlaylist) (FetchResult, error) {
-	switch p.Source {
-	case "apple_music":
-		name, art, tracks, err := fetchAppleMusicPlaylist(p.SourceURL)
-		return FetchResult{name, art, tracks}, err
-	case "spotify":
-		name, art, tracks, err := fetchSpotifyPlaylist(p.SourceURL)
-		return FetchResult{name, art, tracks}, err
-	default: // "listenbrainz" or legacy empty
-		mbid := p.LBMBID
-		if mbid == "" && p.SourceURL != "" {
-			var err error
-			mbid, err = extractLBMBID(p.SourceURL)
-			if err != nil {
-				return FetchResult{}, err
-			}
-		}
-		if mbid == "" {
-			return FetchResult{}, fmt.Errorf("no source data for playlist %s", p.ID)
-		}
-		httpClient := util.NewHttp(util.HttpClientConfig{Timeout: 30})
-		name, modelTracks, err := discovery.FetchPlaylistByMBID(httpClient, mbid)
-		if err != nil {
-			return FetchResult{}, err
-		}
-		tracks := modelTracksToPlaylistTracks(modelTracks)
-		return FetchResult{Name: name, Tracks: tracks}, nil
-	}
-}
-
-// extractSourceID validates a URL and returns the canonical ID for the given source.
-func extractSourceID(source, url string) (string, error) {
-	switch source {
-	case "apple_music":
-		return extractAppleMusicID(url)
-	case "spotify":
-		return extractSpotifyID(url)
-	default:
-		return extractLBMBID(url)
-	}
-}
-
-// isDuplicate checks whether a playlist with the same source and source ID already exists.
-func isDuplicate(source, sourceID string, existing []CustomPlaylist) (string, bool) {
-	for _, p := range existing {
-		if p.Source != source && p.Source != "" {
-			continue
-		}
-		existID, _ := extractSourceID(p.Source, p.SourceURL)
-		if existID == "" && p.LBMBID != "" {
-			existID = p.LBMBID
-		}
-		if existID == "" {
-			continue
-		}
-		if existID == sourceID {
-			return p.ID, true
-		}
-	}
-	return "", false
-}
 
 // handleGetCustomPlaylists returns all saved custom playlists with a track_count
 // derived from their cache file (if present) and the current sync schedule from .env.
-func (s *Server) handleGetCustomPlaylists(w http.ResponseWriter, r *http.Request) {
-	playlists := loadCustomPlaylists(s.cfg.WebDataDir)
+func (p *Playlist) HandleGetCustomPlaylists(w http.ResponseWriter, r *http.Request) {
+	playlists := loadCustomPlaylists(p.cfg.WebDataDir)
 
 	// Read .env to look up schedule state for each custom playlist.
 	var envValues map[string]string
-	if data, err := os.ReadFile(s.cfg.WebEnvPath); err == nil {
-		envValues = parseEnvText(string(data))
+	if data, err := os.ReadFile(p.cfg.WebEnvPath); err == nil {
+		envValues = p.settings.ParseEnvText(string(data))
 	} else {
 		envValues = map[string]string{}
 	}
@@ -228,12 +39,12 @@ func (s *Server) handleGetCustomPlaylists(w http.ResponseWriter, r *http.Request
 		Flags      string `json:"flags"`
 	}
 	items := make([]respItem, 0, len(playlists))
-	for _, p := range playlists {
-		count := customPlaylistTrackCount(s.cfg.WebDataDir, p.ID)
-		prefix := customEnvPrefix(p.Name)
+	for _, plist := range playlists {
+		count := customPlaylistTrackCount(p.cfg.WebDataDir, plist.ID)
+		prefix := util.CustomEnvPrefix(plist.Name)
 		sched := envValues[prefix+"_SCHEDULE"]
 		flags := envValues[prefix+"_FLAGS"]
-		items = append(items, respItem{CustomPlaylist: p, TrackCount: count, Schedule: sched, Flags: flags})
+		items = append(items, respItem{CustomPlaylist: plist, TrackCount: count, Schedule: sched, Flags: flags})
 	}
 
 	w.Header().Set("Content-Type", "application/json")
@@ -244,7 +55,7 @@ func (s *Server) handleGetCustomPlaylists(w http.ResponseWriter, r *http.Request
 
 // handleImportCustomPlaylist imports a playlist by URL (ListenBrainz or Apple Music),
 // writes a cache, and returns the playlist name/tracks to the frontend for the import animation.
-func (s *Server) handleImportCustomPlaylist(w http.ResponseWriter, r *http.Request) {
+func (p *Playlist) HandleImportCustomPlaylist(w http.ResponseWriter, r *http.Request) {
 	var body struct {
 		URL         string `json:"url"`
 		Source      string `json:"source"` // "listenbrainz" | "apple_music"
@@ -255,7 +66,7 @@ func (s *Server) handleImportCustomPlaylist(w http.ResponseWriter, r *http.Reque
 		return
 	}
 
-	existing := loadCustomPlaylists(s.cfg.WebDataDir)
+	existing := loadCustomPlaylists(p.cfg.WebDataDir)
 
 	if body.Source == "" {
 		body.Source = "listenbrainz"
@@ -290,7 +101,7 @@ func (s *Server) handleImportCustomPlaylist(w http.ResponseWriter, r *http.Reque
 	slog.Info("custom-playlists: fetched", "source", body.Source, "name", name, "tracks", len(tracks))
 
 	// Ensure data directories exist before writing anything
-	if err := os.MkdirAll(filepath.Join(s.cfg.WebDataDir, "cache"), 0755); err != nil {
+	if err := os.MkdirAll(filepath.Join(p.cfg.WebDataDir, "cache"), 0755); err != nil {
 		slog.Error("custom-playlists: failed to create data dir", "err", err)
 		http.Error(w, "server data directory unavailable: "+err.Error(), http.StatusInternalServerError)
 		return
@@ -302,18 +113,18 @@ func (s *Server) handleImportCustomPlaylist(w http.ResponseWriter, r *http.Reque
 	// Write cache with remote cover URLs synchronously so the response is fast,
 	// then download local copies of cover art in the background.
 	slog.Info("custom-playlists: writing cache", "id", id)
-	if !writePreliminaryCache(s.cfg.WebDataDir, id, tracks) {
+	if !writePreliminaryCache(p.cfg.WebDataDir, id, tracks) {
 		http.Error(w, "failed to write playlist cache", http.StatusInternalServerError)
 		return
 	}
-	go downloadAndCacheCovers(s.cfg.WebDataDir, id, tracks)
+	go downloadAndCacheCovers(p.cfg.WebDataDir, id, tracks)
 
 	// Cache the playlist's own artwork locally so we can later push it to the
 	// music app on first playlist creation. Apple Music imports have artwork;
 	// ListenBrainz don't.
 	if artworkURL != "" {
 		go func() {
-			if _, err := util.DownloadFile(artworkURL, CustomPlaylistArtworkPath(s.cfg.WebDataDir, id)); err != nil {
+			if _, err := util.DownloadFile(artworkURL, CustomPlaylistArtworkPath(p.cfg.WebDataDir, id)); err != nil {
 				slog.Warn("custom-playlists: artwork download failed", "id", id, "err", err.Error())
 			}
 		}()
@@ -338,7 +149,7 @@ func (s *Server) handleImportCustomPlaylist(w http.ResponseWriter, r *http.Reque
 		LastFetched: time.Now().UTC(),
 	}
 	existing = append(existing, cp)
-	if err := saveCustomPlaylists(s.cfg.WebDataDir, existing); err != nil {
+	if err := saveCustomPlaylists(p.cfg.WebDataDir, existing); err != nil {
 		slog.Error("custom-playlists: failed to save metadata", "err", err)
 		http.Error(w, "failed to save playlist metadata: "+err.Error(), http.StatusInternalServerError)
 		return
@@ -348,14 +159,14 @@ func (s *Server) handleImportCustomPlaylist(w http.ResponseWriter, r *http.Reque
 	// a daily poll SCHEDULE — RefreshDays in the JSON gates the actual refresh interval
 	// inside the cron task body. "Never" imports get FLAGS only so the card is usable
 	// for manual runs while the schedule editor pre-selects "Never".
-	prefix := customEnvPrefix(name)
+	prefix := util.CustomEnvPrefix(name)
 	envUpdates := map[string]string{
 		prefix + "_FLAGS": "--playlist " + id,
 	}
 	if body.RefreshDays > 0 {
 		envUpdates[prefix+"_SCHEDULE"] = "0 4 * * *"
 	}
-	_ = updateEnvKeys(s.cfg.WebEnvPath, envUpdates, web.SampleEnv)
+	_ = p.settings.UpdateEnvKeys(envUpdates, web.SampleEnv)
 
 	slog.Info("custom-playlists: import complete", "id", id, "name", name)
 
@@ -387,14 +198,14 @@ func (s *Server) handleImportCustomPlaylist(w http.ResponseWriter, r *http.Reque
 
 // handleRefreshCustomPlaylist re-fetches a custom playlist and updates the cache.
 // Equivalent to manually triggering the nightly refresh cron job for a single playlist.
-func (s *Server) handleRefreshCustomPlaylist(w http.ResponseWriter, r *http.Request) {
+func (p *Playlist) HandleRefreshCustomPlaylist(w http.ResponseWriter, r *http.Request) {
 	id := r.PathValue("id")
 	if !defs.CustomIDRe.MatchString(id) {
 		http.Error(w, "invalid playlist id", http.StatusBadRequest)
 		return
 	}
 
-	playlists := loadCustomPlaylists(s.cfg.WebDataDir)
+	playlists := loadCustomPlaylists(p.cfg.WebDataDir)
 	idx := -1
 	for i, p := range playlists {
 		if p.ID == id {
@@ -407,10 +218,10 @@ func (s *Server) handleRefreshCustomPlaylist(w http.ResponseWriter, r *http.Requ
 		return
 	}
 
-	p := playlists[idx]
-	slog.Info("custom-playlists: manual refresh", "id", id, "source", p.Source)
+	plist := playlists[idx]
+	slog.Info("custom-playlists: manual refresh", "id", id, "source", plist.Source)
 
-	result, err := fetchCustomPlaylistTracks(p)
+	result, err := fetchCustomPlaylistTracks(plist)
 	if err != nil {
 		slog.Error("custom-playlists: refresh fetch failed", "id", id, "err", err)
 		http.Error(w, "failed to fetch playlist: "+err.Error(), http.StatusBadGateway)
@@ -418,14 +229,14 @@ func (s *Server) handleRefreshCustomPlaylist(w http.ResponseWriter, r *http.Requ
 	}
 	tracks := result.Tracks
 
-	if !writePreliminaryCache(s.cfg.WebDataDir, id, tracks) {
+	if !writePreliminaryCache(p.cfg.WebDataDir, id, tracks) {
 		http.Error(w, "failed to write playlist cache", http.StatusInternalServerError)
 		return
 	}
-	go downloadAndCacheCovers(s.cfg.WebDataDir, id, tracks)
+	go downloadAndCacheCovers(p.cfg.WebDataDir, id, tracks)
 
 	playlists[idx].LastFetched = time.Now().UTC()
-	if err := saveCustomPlaylists(s.cfg.WebDataDir, playlists); err != nil {
+	if err := saveCustomPlaylists(p.cfg.WebDataDir, playlists); err != nil {
 		slog.Warn("custom-playlists: failed to update last_fetched after refresh", "err", err)
 	}
 
@@ -439,7 +250,7 @@ func (s *Server) handleRefreshCustomPlaylist(w http.ResponseWriter, r *http.Requ
 // handleDeleteCustomPlaylist removes a custom playlist's metadata and cache file.
 // If ?delete_tracks=true is set and USE_SUBDIRECTORY is on, also removes the
 // playlist's download subdirectories from DOWNLOAD_DIR.
-func (s *Server) handleDeleteCustomPlaylist(w http.ResponseWriter, r *http.Request) {
+func (p *Playlist) HandleDeleteCustomPlaylist(w http.ResponseWriter, r *http.Request) {
 	id := r.PathValue("id")
 	if !defs.CustomIDRe.MatchString(id) {
 		slog.Warn("custom-playlists: invalid id in delete request", "id", id)
@@ -449,7 +260,7 @@ func (s *Server) handleDeleteCustomPlaylist(w http.ResponseWriter, r *http.Reque
 	deleteTracks := r.URL.Query().Get("delete_tracks") == "true"
 	slog.Info("custom-playlists: delete request", "id", id, "delete_tracks", deleteTracks)
 
-	existing := loadCustomPlaylists(s.cfg.WebDataDir)
+	existing := loadCustomPlaylists(p.cfg.WebDataDir)
 	filtered := existing[:0]
 	found := false
 	var deletedName string
@@ -466,25 +277,25 @@ func (s *Server) handleDeleteCustomPlaylist(w http.ResponseWriter, r *http.Reque
 		return
 	}
 
-	if err := saveCustomPlaylists(s.cfg.WebDataDir, filtered); err != nil {
+	if err := saveCustomPlaylists(p.cfg.WebDataDir, filtered); err != nil {
 		http.Error(w, "failed to save: "+err.Error(), http.StatusInternalServerError)
 		return
 	}
 
 	// Remove the cache file; ignore error if already gone
-	cachePath := filepath.Join(s.cfg.WebDataDir, "cache", id+".json")
+	cachePath := filepath.Join(p.cfg.WebDataDir, "cache", id+".json")
 	_ = os.Remove(cachePath)
 
 	// Remove schedule env vars from .env
-	prefix := customEnvPrefix(deletedName)
-	_ = updateEnvKeys(s.cfg.WebEnvPath, map[string]string{
+	prefix := util.CustomEnvPrefix(deletedName)
+	_ = p.settings.UpdateEnvKeys(map[string]string{
 		prefix + "_SCHEDULE": "",
 		prefix + "_FLAGS":    "",
 	}, web.SampleEnv)
 
 	if deleteTracks {
-		if data, err := os.ReadFile(s.cfg.WebEnvPath); err == nil {
-			env := parseEnvText(string(data))
+		if data, err := os.ReadFile(p.cfg.WebEnvPath); err == nil {
+			env := p.settings.ParseEnvText(string(data))
 			if env["USE_SUBDIRECTORY"] == "true" && env["DOWNLOAD_DIR"] != "" {
 				prefix := cases.Title(language.Und).String(id) // "custom-1234" -> "Custom-1234"
 				removed, err := util.RemoveDirsByPrefix(env["DOWNLOAD_DIR"], prefix)
@@ -493,7 +304,7 @@ func (s *Server) handleDeleteCustomPlaylist(w http.ResponseWriter, r *http.Reque
 				} else {
 					slog.Info("custom-playlists: removed download dirs", "id", id, "count", removed)
 					if removed > 0 {
-						s.triggerLibraryRefresh()
+						util.TriggerLibraryRefresh(p.cfg)
 					}
 				}
 			} else {
@@ -507,19 +318,152 @@ func (s *Server) handleDeleteCustomPlaylist(w http.ResponseWriter, r *http.Reque
 	w.WriteHeader(http.StatusNoContent)
 }
 
-// customPlaylistTrackCount reads the cached track count for a custom playlist without
-// fully parsing the JSON.
-func customPlaylistTrackCount(cfgDir, id string) int {
-	type mini struct {
-		Tracks []json.RawMessage `json:"tracks"`
+// handleSaveSchedule updates a single playlist's schedule in the .env file.
+func (p *Playlist) HandleSaveSchedule(w http.ResponseWriter, r *http.Request) {
+	var body struct {
+		Name    string `json:"name"`
+		Enabled bool   `json:"enabled"`
+		Day     int    `json:"day"` // 0=Sun…6=Sat, -1=every day
+		Hour    int    `json:"hour"`
+		Minute  int    `json:"minute"`
 	}
-	data, err := os.ReadFile(filepath.Join(cfgDir, "cache", id+".json"))
-	if err != nil {
-		return 0
+	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
+		http.Error(w, "invalid JSON: "+err.Error(), http.StatusBadRequest)
+		return
 	}
-	var m mini
-	if err := json.Unmarshal(data, &m); err != nil {
-		return 0
+
+	var envPrefix string
+	var defaultFlags string
+
+	if def, ok := defs.PlaylistDefs[body.Name]; ok {
+		envPrefix = def.EnvPrefix
+		defaultFlags = def.DefaultFlags
+	} else if defs.CustomIDRe.MatchString(body.Name) {
+		envPrefix = util.CustomEnvPrefix(body.Name)
+		defaultFlags = "--playlist " + body.Name
+	} else {
+		http.Error(w, "unknown playlist name", http.StatusBadRequest)
+		return
 	}
-	return len(m.Tracks)
+
+	updates := map[string]string{}
+	if !body.Enabled {
+		// Toggle off — truly disable, regardless of day value carried over from state
+		updates[envPrefix+"_SCHEDULE"] = ""
+		updates[envPrefix+"_FLAGS"] = ""
+	} else if body.Day == -2 {
+		// "Never" — keep playlist active for manual runs but remove auto-schedule
+		updates[envPrefix+"_SCHEDULE"] = ""
+		updates[envPrefix+"_FLAGS"] = defaultFlags
+	} else {
+		dom := "*"
+		dow := "*"
+		if body.Day == 100 {
+			dom = "1"
+		} else if body.Day >= 0 {
+			dow = fmt.Sprintf("%d", body.Day)
+		}
+		updates[envPrefix+"_SCHEDULE"] = fmt.Sprintf("%d %d %s * %s", body.Minute, body.Hour, dom, dow)
+		updates[envPrefix+"_FLAGS"] = defaultFlags
+	}
+
+	if err := p.settings.UpdateEnvKeys(updates, web.SampleEnv); err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	w.WriteHeader(http.StatusOK)
+}
+
+// handleGetPlaylist serves the tracklist cache written by explo during its last run.
+// Returns an empty track list if no cache exists yet.
+func (p *Playlist) HandleGetPlaylist(w http.ResponseWriter, r *http.Request) {
+	playlistType := r.URL.Query().Get("type")
+	if !isValidPlaylistID(playlistType) {
+		http.Error(w, "unknown playlist type", http.StatusBadRequest)
+		return
+	}
+
+	cachePath := filepath.Join(p.cfg.WebDataDir, "cache", playlistType+".json")
+	if raw, err := os.ReadFile(cachePath); err == nil {
+		w.Header().Set("Content-Type", "application/json")
+		if _, err := w.Write(raw); err != nil {
+			slog.Error("failed to write playlist response", "msg", err.Error())
+		}
+		return
+	}
+
+	// No cache yet — return an empty response. Run explo or use the prefetch
+	// endpoint to populate the cache.
+	w.Header().Set("Content-Type", "application/json")
+	if _, err := w.Write([]byte(`{"tracks":[]}`)); err != nil {
+		slog.Error("failed to write empty playlist response", "msg", err.Error())
+	}
+}
+
+// handlePrefetchCovers fetches the most recent LB playlists for the given user,
+// writes a preliminary JSON cache for the web UI, then downloads cover art.
+// Runs in the background — returns 202 immediately.
+func (p *Playlist) HandlePrefetchCovers(w http.ResponseWriter, r *http.Request) {
+	var body struct {
+		User      string   `json:"user"`
+		Playlists []string `json:"playlists"`
+		Source    string   `json:"source"`
+	}
+	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
+		http.Error(w, "invalid JSON: "+err.Error(), http.StatusBadRequest)
+		return
+	}
+	if body.User == "" || len(body.Playlists) == 0 {
+		http.Error(w, "user and playlists are required", http.StatusBadRequest)
+		return
+	}
+	forceRefresh := body.Source == "wizard"
+	w.WriteHeader(http.StatusAccepted)
+
+	slog.Info("prefetch: starting", "user", body.User, "playlists", body.Playlists, "source", body.Source, "force_refresh", forceRefresh)
+	go func() {
+		for _, pt := range body.Playlists {
+			if !validPlaylistTypes[pt] {
+				slog.Warn("prefetch: unknown playlist type", "type", pt)
+				continue
+			}
+			// Normal prefetch keeps an existing cache intact; wizard prefetch refreshes it
+			// after the user updates discovery settings.
+			cachePath := filepath.Join(p.cfg.WebDataDir, "cache", pt+".json")
+			if _, err := os.Stat(cachePath); err == nil && !forceRefresh {
+				slog.Info("prefetch: cache already exists, skipping", "playlist", pt)
+				continue
+			}
+			var tracks []PlaylistTrack
+			var err error
+			if pt == "on-repeat" {
+				tracks, err = fetchOnRepeatTracks(body.User)
+			} else {
+				tracks, err = fetchMostRecentLBPlaylist(body.User, pt)
+			}
+			if err != nil {
+				slog.Warn("prefetch: failed to fetch LB playlist", "type", pt, "err", err)
+				continue
+			}
+			slog.Info("prefetch: fetched tracks", "playlist", pt, "count", len(tracks))
+			writePrefetchCache(p.cfg.WebDataDir, pt, tracks)
+		}
+	}()
+}
+
+// handleBackgroundArt returns a single cover art URL for use as a login page backdrop.
+// It picks a random local cover if any exist; otherwise it fetches the top global
+// albums from ListenBrainz and downloads cover art for the first available one.
+func (p *Playlist) HandleBackgroundArt(w http.ResponseWriter, r *http.Request) {
+	coversDir := filepath.Join(p.cfg.WebDataDir, "cache", "covers")
+
+	url := randomLocalCoverHiRes(coversDir)
+	if url == "" {
+		url = fetchSitewideCovers(coversDir)
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+	if err := json.NewEncoder(w).Encode(map[string]string{"url": url}); err != nil {
+		slog.Error("background-art: failed to write response", "err", err.Error())
+	}
 }
