+++ src/web/backend/playlist/spotify.go
@@ -1,4 +1,4 @@
-package backend
+package playlist
 
 import (
 	"bytes"
